moving _parse_docstring to hca.utils._doc
