+++ hca/util/_docs.py
@@ -1,5 +1,12 @@
 import CommonMark
 
+from docutils import utils
+from docutils.frontend import OptionParser
+from docutils.parsers.rst import Parser as RSTParser_component
+from sphinx.parsers import RSTParser
+
+from .. import logger
+
 _pagination_docstring = """
 .. admonition:: Pagination
 
@@ -40,3 +47,53 @@ def _md2rst(docstring):
     ast = parser.parse(docstring)
     renderer = CommonMark.ReStructuredTextRenderer()
     return renderer.render(ast)
+
+
+def _parse_docstring(docstring):
+    """
+    Using the sphinx RSTParse to parse __doc__ for argparse `parameters`, `help`, and `description`. The first
+    rst paragraph encountered it treated as the argparse help text. Any param fields are treated as argparse
+    arguments. Any other text is combined and added to the argparse description.
+
+    example:
+        \"""
+         this will be the summary
+
+         :param name: describe the parameter called name.
+
+         this will be the descriptions
+
+         * more description
+         * more description
+
+         This will also be in the description
+        \"""
+
+    :param str docstring:
+    :return:
+    :rtype: dict
+    """
+    settings = OptionParser(components=(RSTParser_component,)).get_default_values()
+    rstparser = RSTParser()
+    document = utils.new_document(' ', settings)
+    rstparser.parse(docstring, document)
+    if document.children[0].tagname != 'block_quote':
+        logger.warning("The first line of the docstring must be blank.")
+    else:
+        document = document.children[0]
+
+    def get_params(field_list_node, params):
+        for field in field_list_node.children:
+            name = field.children[0].rawsource.split(' ')
+            if 'param' == name[0]:
+                params[name[-1]] = field.children[1].astext()
+
+    method_args = {'summary': '', 'params': dict(), 'description': ''}
+    for node in document.children:
+        if node.tagname is 'paragraph' and method_args['summary'] == '':
+            method_args['summary'] = node.astext()
+        elif node.tagname is 'field_list':
+            get_params(node, method_args['params'])
+        else:
+            method_args['description'] += '\n' + node.astext()
+    return method_args
